pyproject.toml
[project]
name = "article-backup"
version = "0.3.5"
description = "Локальный бэкап статей с Sponsr.ru и Boosty.to в Markdown с Hugo-интеграцией"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Eugene Chaykin", email = "eugene@chayk.in"}
]
requires-python = ">=3.10"
keywords = ["backup", "markdown", "sponsr", "boosty", "archiver", "hugo"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: System :: Archiving :: Backup",
]
dependencies = [
    "requests",
    "beautifulsoup4",
    "lxml",
    "html2text",
    "python-slugify",
    "pyyaml",
]

[project.urls]
Homepage = "https://github.com/strannick-ru/article-backup"
Repository = "https://github.com/strannick-ru/article-backup"
"Bug Tracker" = "https://github.com/strannick-ru/article-backup/issues"
Changelog = "https://github.com/strannick-ru/article-backup/blob/main/CHANGELOG.md"

[project.scripts]
article-backup = "backup:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["backup"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
